# dummy_api/ver3/dummy_api_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IoT Monitoring API",
    "description": "API for IoT device data (simulated sensors & meters)",
    "version": "1.0.0"
  },
  "paths": {
    "/temperature": {
      "get": {
        "summary": "Get latest temperature reading with environment and internal temps",
        "operationId": "get_temperature",
        "responses": {
          "200": {
            "description": "Temperature data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemperatureReading"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Submit new environment and internal temperatures",
        "operationId": "post_temperature",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemperatureReading"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Temperature data received",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemperatureReading"
                }
              }
            }
          }
        }
      }
    },
    "/energy": {
      "get": {
        "summary": "Get latest energy meter reading",
        "operationId": "get_energy",
        "responses": {
          "200": {
            "description": "Energy meter data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnergyReading"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Submit new energy meter reading",
        "operationId": "post_energy",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnergyReading"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Energy data received",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnergyReading"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EnvironmentTemperature": {
        "type": "object",
        "properties": {
          "temp_c": {
            "type": "number",
            "example": 22.5
          }
        }
      },
      "InternalTemperature": {
        "type": "object",
        "properties": {
          "temp_c": {
            "type": "number",
            "example": 45.0
          }
        }
      },
      "TemperatureReading": {
        "type": "object",
        "properties": {
          "device_id": {
            "type": "string",
            "example": "sensor-123"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2025-03-19T12:00:00Z"
          },
          "environment": {
            "$ref": "#/components/schemas/EnvironmentTemperature"
          },
          "internal": {
            "$ref": "#/components/schemas/InternalTemperature"
          }
        }
      },
      "EnergyReading": {
        "type": "object",
        "properties": {
          "device_id": {
            "type": "string",
            "example": "meter-456"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2025-03-19T12:00:00Z"
          },
          "power_kw": {
            "type": "number",
            "example": 1.2
          },
          "energy_kwh": {
            "type": "number",
            "example": 500.5
          }
        }
      }
    }
  }
}